+++ Sources/Core/Caching/DataCache.swift
@@ -4,32 +4,6 @@
 
 import Foundation
 
-// MARK: - DataCaching
-
-/// Data cache.
-///
-/// - warning: The implementation must be thread safe.
-public protocol DataCaching {
-    /// Retrieves data from cache for the given key.
-    func cachedData(for key: String) -> Data?
-
-    /// Returns `true` if the cache contains data for the given key.
-    func containsData(for key: String) -> Bool
-
-    /// Stores data for the given key.
-    /// - note: The implementation must return immediately and store data
-    /// asynchronously.
-    func storeData(_ data: Data, for key: String)
-
-    /// Removes data for the given key.
-    func removeData(for key: String)
-
-    /// Removes all items.
-    func removeAll()
-}
-
-// MARK: - DataCache
-
 /// An LRU disk cache that stores data in separate files.
 ///
 /// The DataCache uses LRU cleanup policy (least recently used items are removed
