+++ Sources/Core/Caching/ImagePipelineCache.swift
