+++ Sources/Core/Loading/DataLoading.swift
@@ -1,0 +1,23 @@
+// The MIT License (MIT)
+//
+// Copyright (c) 2015-2021 Alexander Grebenyuk (github.com/kean).
+
+import Foundation
+
+/// A unit of work that can be cancelled.
+public protocol Cancellable: AnyObject {
+    func cancel()
+}
+
+/// Fetches original image data.
+public protocol DataLoading {
+    /// - parameter didReceiveData: Can be called multiple times if streaming
+    /// is supported.
+    /// - parameter completion: Must be called once after all (or none in case
+    /// of an error) `didReceiveData` closures have been called.
+    func loadData(with request: URLRequest,
+                  didReceiveData: @escaping (Data, URLResponse) -> Void,
+                  completion: @escaping (Error?) -> Void) -> Cancellable
+}
+
+extension URLSessionTask: Cancellable {}
